Properties cleanup - part 2
